+++ rust/main/lander/src/adapter/chains/ethereum/nonce.rs
@@ -1,6 +1,4 @@
-pub use manager::NonceManager;
-pub use state::NonceManagerState;
-pub use updater::NonceUpdater;
+pub(crate) use manager::NonceManager;
 
 pub mod db;
 mod error;
@@ -8,6 +6,10 @@ mod manager;
 mod state;
 mod status;
 mod updater;
+
+#[cfg(test)]
+pub(crate) use state::NonceManagerState;
+pub(crate) use updater::NonceUpdater;
 
 #[cfg(test)]
 mod tests;
