+++ rust/main/lander/src/adapter/chains/ethereum/nonce/state/db.rs
@@ -3,7 +3,7 @@ use hyperlane_core::U256;
 use crate::transaction::{Transaction, TransactionUuid};
 
 use super::super::error::NonceResult;
-use super::super::NonceManagerState;
+use super::NonceManagerState;
 
 impl NonceManagerState {
     pub(super) async fn get_boundary_nonces(&self) -> NonceResult<(Option<U256>, U256)> {
